requirements.txt
numpy
torch
Pillow
numba
